requirements.txt
feedparser==6.0.11
google-generativeai==0.8.2
requests==2.32.3